global: implement and refactor developer workflow
