+++ invenio_cli/cli/cli.py
@@ -13,10 +13,15 @@ from pathlib import Path
 
 import click
 
-from .commands import Commands, ContainersCommands, LocalCommands
-from .errors import InvenioCLIConfigError
-from .helpers.cli_config import CLIConfig
-from .helpers.cookiecutter_wrapper import CookiecutterWrapper
+from ..commands import Commands, ContainersCommands, InstallCommands, \
+    LocalCommands
+from ..errors import InvenioCLIConfigError
+from ..helpers.cli_config import CLIConfig
+from ..helpers.cookiecutter_wrapper import CookiecutterWrapper
+from .assets import assets
+from .packages import packages
+from .services import services
+
 
 pass_cli_config = click.make_pass_decorator(CLIConfig, ensure=True)
 
@@ -24,7 +29,7 @@ pass_cli_config = click.make_pass_decorator(CLIConfig, ensure=True)
 @click.group()
 @click.version_option()
 @click.pass_context
-def cli(ctx):
+def invenio_cli(ctx):
     """Initialize CLI context."""
     # Config loading is not needed when initializing
     if ctx.invoked_subcommand != "init":
@@ -34,7 +39,11 @@ def cli(ctx):
             click.secho(e.message, fg="red")
 
 
-@cli.command()
+invenio_cli.add_command(assets)
+invenio_cli.add_command(packages)
+invenio_cli.add_command(services)
+
+@invenio_cli.command()
 @click.argument('flavour', type=click.Choice(['RDM'], case_sensitive=False),
                 default='RDM', required=False)
 @click.option('-t', '--template', required=False,
@@ -53,7 +62,7 @@ def init(flavour, template, checkout):
         click.secho("Running cookiecutter...", fg='green')
         project_dir = cookiecutter_wrapper.cookiecutter()
 
-        click.secho("Writing invenio-cli config file...", fg='green')
+        click.secho("Writing invenio-invenio_cli config file...", fg='green')
         saved_replay = cookiecutter_wrapper.get_replay()
         CLIConfig.write(project_dir, flavour, saved_replay)
 
@@ -67,73 +76,40 @@ def init(flavour, template, checkout):
         cookiecutter_wrapper.remove_config()
 
 
-@cli.command()
+@invenio_cli.command()
 @click.option('--pre', default=False, is_flag=True,
               help='If specified, allows the installation of alpha releases')
-@click.option('--lock/--skip-lock', default=True, is_flag=True,
-              help='Lock dependencies or avoid this step')
 @click.option(
     '--production/--development', '-p/-d', default=True, is_flag=True,
     help='Production mode copies statics/assets. Development mode symlinks'
          ' statics/assets.'
 )
 @pass_cli_config
-def install(cli_config, pre, lock, production):
+def install(cli_config, pre, production):
     """Installs the  project locally.
 
     Installs dependencies, creates instance directory,
     links invenio.cfg + templates, copies images and other statics and finally
     builds front-end assets.
     """
-    commands = LocalCommands(cli_config)
-    commands.install(
+    commands = InstallCommands(cli_config)
+    click.secho()
+
+    result = commands.install(
         pre=pre,
-        lock=lock,
         flask_env='production' if production else 'development'
     )
 
-
-@cli.command()
-@click.option('-f', '--force', default=False, is_flag=True,
-              help='Force recreation of db tables, ES indices, queues...')
-@pass_cli_config
-def services(cli_config, force):
-    """Starts DB, ES, queue and cache services and ensures they are setup.
-
-    --force destroys and resets services
-    """
-    commands = LocalCommands(cli_config)
-    commands.services(force=force)
-
-
-@cli.command()
-@click.option('-v', '--verbose', default=False, is_flag=True, required=False,
-              help='Verbose mode will show all logs in the console.')
-@pass_cli_config
-def status(cli_config, verbose):
-    """Checks if the services are up and running.
-
-    NOTE: currently only ES, DB (postgresql/mysql) and redis are supported.
-    """
-    commands = Commands(cli_config)
-    services = ["redis", cli_config.get_db_type(), "es"]
-    statuses = commands.status(services=services, verbose=verbose)
-
-    messages = [
-        {"message": "{}: up and running.", "fg": "green"},
-        {"message": "{}: unable to connect or bad response.", "fg": "red"},
-        {"message": "{}: no healthcheck function defined.", "fg": "yellow"}
-    ]
-
-    for idx, status in enumerate(statuses):
-        message = messages[status]
+    if result.status_code > 0:
         click.secho(
-            message=message.get("message").format(services[idx]),
-            fg=message.get("fg")
+            "Failed to install dependencies.\nErrors: {}\nOutput: {}",
+            fg="red"
         )
+    else:
+        click.secho("Dependencies installed successfully.", fg="green")
 
 
-@cli.command()
+@invenio_cli.command()
 @click.option('--pre', default=False, is_flag=True,
               help='If specified, allows the installation of alpha releases')
 @click.option('-f', '--force', default=False, is_flag=True,
@@ -151,17 +127,7 @@ def containerize(cli_config, pre, force, install_js):
     commands.containerize(pre=pre, force=force, install=install_js)
 
 
-@cli.command()
-@click.option('--local/--containers', default=True, is_flag=True,
-              help='Which environment to build, it defaults to local')
-@pass_cli_config
-def demo(cli_config, local):
-    """Populates instance with demo records."""
-    commands = LocalCommands(cli_config)
-    commands.demo()
-
-
-@cli.command()
+@invenio_cli.command()
 @click.option('--host', '-h',  default='127.0.0.1',
               help='The interface to bind to.')
 @click.option('--port', '-p',  default=5000,
@@ -169,67 +135,19 @@ def demo(cli_config, local):
 @click.option('--debug/--no-debug', '-d/',  default=True, is_flag=True,
               help='Enable/disable debug mode including auto-reloading '
                    '(default: enabled).')
+@click.option('--services/--no-services', '-s/-n',  default=True, is_flag=True,
+              help='Enable/disable dockerized services (default: enabled).')
 @pass_cli_config
-def run(cli_config, host, port, debug):
+def run(cli_config, host, port, debug, services):
     """Starts the local development server.
 
     NOTE: this only makes sense locally so no --local option
     """
     commands = LocalCommands(cli_config)
-    commands.run(host=host, port=str(port), debug=debug)
+    commands.run(host=host, port=str(port), debug=debug, services=services)
 
 
-@cli.group()
-def assets():
-    """Statics and assets management commands."""
-
-
-@assets.command()
-@click.option('--force', '-f', default=False, is_flag=True,
-              help='Force the full recreation the assets and statics.')
-@click.option(
-    '--production/--development', '-p/-d', default=True, is_flag=True,
-    help='Production mode copies files. Development mode symlinks files.'
-)
-@pass_cli_config
-def update(cli_config, force, production):
-    """Updates the current application static/assets files."""
-    commands = LocalCommands(cli_config)
-    commands.update_statics_and_assets(
-        force=force,
-        flask_env='production' if production else 'development'
-    )
-
-
-@assets.command()
-@pass_cli_config
-def watch(cli_config):
-    """Watch assets files for changes and rebuild."""
-    commands = LocalCommands(cli_config)
-    commands.watch_assets()
-
-
-@assets.command('install-module')
-@click.argument('path', type=click.Path(exists=True))
-@pass_cli_config
-def install_module(cli_config, path):
-    """Install and link a React module."""
-    commands = LocalCommands(cli_config)
-    commands.link_js_module(path)
-
-
-@assets.command('watch-module')
-@click.option('--link', '-l', default=False, is_flag=True,
-              help='Link the module.')
-@click.argument('path', type=click.Path(exists=True))
-@pass_cli_config
-def watch_module(cli_config, path, link):
-    """Watch a React module."""
-    commands = LocalCommands(cli_config)
-    commands.watch_js_module(path, link=link)
-
-
-@cli.command()
+@invenio_cli.command()
 @click.option('-v', '--verbose', default=False, is_flag=True, required=False,
               help='Verbose mode will show all logs in the console.')
 @pass_cli_config
@@ -242,17 +160,7 @@ def destroy(cli_config, verbose):
     click.secho('Instance destroyed', fg='green')
 
 
-@cli.command()
-@pass_cli_config
-def stop(cli_config):
-    """Stops containers."""
-    commands = Commands(cli_config)
-    click.secho("Stopping containers...", fg="green")
-    commands.stop()
-    click.secho('Stopped containers', fg='green')
-
-
-@cli.command()
+@invenio_cli.command()
 @click.option('-v', '--verbose', default=False, is_flag=True, required=False,
               help='Verbose mode will show all logs in the console.')
 def upgrade(verbose):
@@ -260,7 +168,7 @@ def upgrade(verbose):
     click.secho('TODO: Implement upgrade command', fg='red')
 
 
-@cli.command()
+@invenio_cli.command()
 @pass_cli_config
 def shell(cli_config, ):
     """Shell command."""
@@ -268,7 +176,7 @@ def shell(cli_config, ):
     commands.shell()
 
 
-@cli.command()
+@invenio_cli.command()
 @click.option(
     '--debug/--no-debug', '-d/', default=False, is_flag=True,
     help='Enable Flask development mode (default: disabled).'
@@ -278,17 +186,3 @@ def pyshell(cli_config, debug):
     """Python shell command."""
     commands = Commands(cli_config)
     commands.pyshell(debug=debug)
-
-
-@cli.group()
-def ext():
-    """Commands for development."""
-
-
-@ext.command('module-install')
-@click.argument("modules", nargs=-1, type=str)
-@pass_cli_config
-def module_install(cli_config, modules):
-    """Install a Python module in development mode."""
-    commands = LocalCommands(cli_config)
-    commands.install_modules(modules)
