+++ setup.py
@@ -65,7 +65,7 @@ setup(
     platforms='any',
     entry_points={
         'console_scripts': [
-            'invenio-cli = invenio_cli.cli:cli',
+            'invenio-cli = invenio_cli.cli:invenio_cli',
         ]
     },
     extras_require=extras_require,
